Fix bostream nonces
Too stupid to notice the difference between bits and bytes, though I
could fit a 24 byte buffer in a 32 bit number (we'd need 192 bits, lack
of big ints strikes again).

Convert the nonces back to an array we can mutate in-place. Disadvantage
is that we now need to have a copy that can be mutated (ref/iso/trn),
and at the same time give it away for sending. Still thinking of how to
do this without allocating every time.